music: rename class to MusicManager and add Music class
