+++ src/YoutubePlaylists.ts
@@ -1,12 +1,10 @@
 import searchApi from "youtube-search-api";
+import Music from "./Music";
 
-async function GetPlaylist(id: string) {
+async function GetPlaylist(id: string): Promise<Music[]> {
     const result = await searchApi.GetPlaylistData(id, 100);
     return result.items.map((item) => {
-        return {
-            name: item.title,
-            id: item.id,
-        };
+        return new Music(item.title, item.id);
     });
 }
